major changes
Fully debugged code so that it runs with HARK 0.16.0, and pinned this to "environment.yml". From now on, when the virtual env is created to run all of this code, it will not depend on the latest HARK@mater release. Lastly, reran vital versions of the code for SCF waves 2004-2019